Moved Trip model to models/Types
